refactor: Remove old builders and consolidate into unified NuruAppBuilder
- Removed DirectAppBuilder and AppBuilder
- Created unified NuruAppBuilder supporting both DI and non-DI modes
- Updated all samples and tests to use new builder
- Updated documentation to reflect the new builder API
- NuruAppBuilder.AddDependencyInjection() enables DI support

This consolidation simplifies the API surface while maintaining the flexibility
to use direct delegates or Mediator commands as needed.
