+++ Source/TimeWarp.Nuru/NuruApp.cs
@@ -1,22 +1,251 @@
 namespace TimeWarp.Nuru;
 
 /// <summary>
-/// A built TimeWarp.Console application.
+/// A unified CLI app that supports both direct execution and dependency injection.
 /// </summary>
 public class NuruApp
 {
-  private readonly IServiceProvider ServiceProvider;
+  private readonly IServiceProvider? ServiceProvider;
+  private readonly EndpointCollection Endpoints;
+  private readonly ITypeConverterRegistry TypeConverterRegistry;
+  private readonly RouteBasedCommandResolver Resolver;
+  private readonly CommandExecutor? CommandExecutor;
 
-  public IServiceProvider Services => ServiceProvider;
-
-  public NuruApp(IServiceProvider serviceProvider)
+  /// <summary>
+  /// Direct constructor - no dependency injection.
+  /// </summary>
+  public NuruApp(EndpointCollection endpoints, ITypeConverterRegistry typeConverterRegistry)
   {
-    ServiceProvider = serviceProvider;
+    Endpoints = endpoints ?? throw new ArgumentNullException(nameof(endpoints));
+    TypeConverterRegistry = typeConverterRegistry ?? throw new ArgumentNullException(nameof(typeConverterRegistry));
+    Resolver = new RouteBasedCommandResolver(endpoints, typeConverterRegistry);
   }
 
-  public Task<int> RunAsync(string[] args)
+  /// <summary>
+  /// DI constructor - with service provider for Mediator support.
+  /// </summary>
+  public NuruApp(IServiceProvider serviceProvider)
   {
-    NuruCli cli = ServiceProvider.GetRequiredService<NuruCli>();
-    return cli.RunAsync(args);
+    ServiceProvider = serviceProvider ?? throw new ArgumentNullException(nameof(serviceProvider));
+    Endpoints = serviceProvider.GetRequiredService<EndpointCollection>();
+    TypeConverterRegistry = serviceProvider.GetRequiredService<ITypeConverterRegistry>();
+    CommandExecutor = serviceProvider.GetRequiredService<CommandExecutor>();
+    Resolver = new RouteBasedCommandResolver(Endpoints, TypeConverterRegistry);
+  }
+
+  public IServiceProvider? Services => ServiceProvider;
+
+  public async Task<int> RunAsync(string[] args)
+  {
+    ArgumentNullException.ThrowIfNull(args);
+
+    try
+    {
+      // Parse and match route
+      ResolverResult result = Resolver.Resolve(args);
+
+      if (!result.Success || result.MatchedEndpoint is null)
+      {
+        await Console.Error.WriteLineAsync(
+          result.ErrorMessage ?? "No matching command found."
+        ).ConfigureAwait(false);
+
+        ShowAvailableCommands();
+        return 1;
+      }
+
+      // Check if this is a Mediator command (only if DI is available)
+      Type? commandType = result.MatchedEndpoint.CommandType;
+
+      if (commandType is not null && ServiceProvider is not null && IsMediatorCommand(commandType))
+      {
+        // Execute through Mediator
+        return await ExecuteMediatorCommandAsync(commandType, result).ConfigureAwait(false);
+      }
+
+      // Execute as delegate
+      if (result.MatchedEndpoint.Handler is Delegate del)
+      {
+        return await ExecuteDelegateAsync(del, result.ExtractedValues).ConfigureAwait(false);
+      }
+
+      await Console.Error.WriteLineAsync(
+        "No valid handler found for the matched route."
+      ).ConfigureAwait(false);
+
+      return 1;
+    }
+    catch (Exception ex)
+    {
+      await Console.Error.WriteLineAsync($"Error: {ex.Message}").ConfigureAwait(false);
+      return 1;
+    }
+  }
+
+  private async Task<int> ExecuteMediatorCommandAsync(Type commandType, ResolverResult result)
+  {
+    if (CommandExecutor is null)
+    {
+      throw new InvalidOperationException("CommandExecutor is not available. Ensure DI is configured.");
+    }
+
+    object? returnValue = await CommandExecutor.ExecuteCommandAsync(
+      commandType,
+      result.ExtractedValues,
+      CancellationToken.None
+    ).ConfigureAwait(false);
+
+    // Handle int return values
+    if (returnValue is int exitCode)
+    {
+      return exitCode;
+    }
+
+    return 0;
+  }
+
+  private async Task<int> ExecuteDelegateAsync(Delegate del, Dictionary<string, string> extractedValues)
+  {
+    try
+    {
+      object?[] args = ServiceProvider is not null
+        ? BindParametersWithDI(del.Method, extractedValues)
+        : BindParameters(del.Method, extractedValues);
+
+      object? returnValue = del.DynamicInvoke(args);
+
+      // Handle async delegates
+      if (returnValue is Task task)
+      {
+        await task.ConfigureAwait(false);
+      }
+
+      return 0;
+    }
+    catch (Exception ex)
+    {
+      await Console.Error.WriteLineAsync(
+        $"Error executing handler: {ex.Message}"
+      ).ConfigureAwait(false);
+
+      return 1;
+    }
+  }
+
+  private object?[] BindParameters(MethodInfo method, Dictionary<string, string> extractedValues)
+  {
+    ParameterInfo[] parameters = method.GetParameters();
+    object?[] args = new object?[parameters.Length];
+
+    for (int i = 0; i < parameters.Length; i++)
+    {
+      ParameterInfo param = parameters[i];
+
+      if (extractedValues.TryGetValue(param.Name!, out string? stringValue))
+      {
+        // Simple type conversion only - no DI
+        if (TypeConverterRegistry.TryConvert(
+          stringValue,
+          param.ParameterType,
+          out object? convertedValue
+        ))
+        {
+          args[i] = convertedValue;
+        }
+        else if (param.ParameterType == typeof(string))
+        {
+          args[i] = stringValue;
+        }
+        else if (param.ParameterType == typeof(string[]))
+        {
+          // Split space-delimited string into array for catch-all parameters
+          args[i] = stringValue.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+        }
+        else
+        {
+          throw new InvalidOperationException(
+            $"Cannot convert '{stringValue}' to type {param.ParameterType} for parameter '{param.Name}'"
+          );
+        }
+      }
+      else if (param.HasDefaultValue)
+      {
+        args[i] = param.DefaultValue;
+      }
+      else
+      {
+        throw new InvalidOperationException(
+          $"No value provided for required parameter '{param.Name}'"
+        );
+      }
+    }
+
+    return args;
+  }
+
+  private object?[] BindParametersWithDI(MethodInfo method, Dictionary<string, string> extractedValues)
+  {
+    // Use DelegateParameterBinder when DI is available
+    // This handles DI injection for parameters
+    ParameterInfo[] parameters = method.GetParameters();
+    object?[] args = new object?[parameters.Length];
+
+    for (int i = 0; i < parameters.Length; i++)
+    {
+      ParameterInfo param = parameters[i];
+
+      // First try extracted values
+      if (extractedValues.TryGetValue(param.Name!, out string? stringValue))
+      {
+        if (TypeConverterRegistry.TryConvert(
+          stringValue,
+          param.ParameterType,
+          out object? convertedValue
+        ))
+        {
+          args[i] = convertedValue;
+        }
+        else if (param.ParameterType == typeof(string))
+        {
+          args[i] = stringValue;
+        }
+        else if (param.ParameterType == typeof(string[]))
+        {
+          args[i] = stringValue.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+        }
+      }
+      // Then try DI for non-value parameters
+      else if (ServiceProvider!.GetService(param.ParameterType) is object service)
+      {
+        args[i] = service;
+      }
+      else if (param.HasDefaultValue)
+      {
+        args[i] = param.DefaultValue;
+      }
+      else
+      {
+        throw new InvalidOperationException(
+          $"No value provided for required parameter '{param.Name}'"
+        );
+      }
+    }
+
+    return args;
+  }
+
+  private static bool IsMediatorCommand(Type type)
+  {
+    return type.GetInterfaces().Any(i =>
+      i.IsGenericType &&
+      (i.GetGenericTypeDefinition() == typeof(IRequest<>) ||
+       i.GetGenericTypeDefinition() == typeof(IRequest))
+    );
+  }
+
+  private void ShowAvailableCommands()
+  {
+    var helpProvider = new Help.RouteHelpProvider(Endpoints);
+    helpProvider.ShowHelp();
   }
 }
